requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
